MIGRATIONS - saving changes before end of night
Made some changes to the users table and added a user_types and
user_permissions table defining the relationship between them in the User
model. Perhaps in my last design I was over thinking it when designing it
around the fact that a user can have many roles however after giving it some
deep thought perhaps this will not be the case and just to keep it simple. A
user has one type and one permission. Going to setup commands to manually be
able to enter these values next